requirements.txt
Flask
Pillow
gunicorn
google-cloud-storage
qrcode[pil]
